requirements.txt
torch==2.4.1
torchvision==0.19.1
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.3.2
matplotlib==3.3.0
tqdm==4.48.2
pillow==7.2.0 